chore: model treeShake
