+++ packages/platform-web/src/node/features/model/index.ts
@@ -20,9 +20,6 @@ const core = createPlugin({
 
 export default {
   core,
-  runtime: {
-    plugin: resolvePkgFile('lib/node/features/model/runtime.js')
-  },
   server,
   types: resolvePkgFile('lib/node/features/model/shuvi-app.d.ts')
 };
