+++ packages/platform-web/src/shuvi-app/react/model/runtime.ts
@@ -1,5 +1,9 @@
-import { Doura } from 'doura';
-import { createRuntimePluginAfter } from '@shuvi/platform-shared/shared';
+import { Doura, doura } from 'doura';
+import {
+  getPageData,
+  IAppContext,
+  createRuntimePluginAfter
+} from '@shuvi/platform-shared/shared';
 import { DouraWrapper } from './Wrapper';
 
 declare module '@shuvi/runtime' {
@@ -8,9 +12,46 @@ declare module '@shuvi/runtime' {
   }
 }
 
+type InitDoura = (params: { initialState: any; ctx: IAppContext }) => Doura;
+
+let currentStore: Doura;
+
+const isServer = typeof window === 'undefined';
+
+// for client, singleton mode
+// for server, return new store
+const initStore: InitDoura = ({ initialState, ctx }) => {
+  const createStoreInstance = () => {
+    return doura({
+      initialState
+    });
+  };
+  // for server
+  if (isServer) {
+    return createStoreInstance();
+  }
+
+  // for client is singleton, just init once
+  if (!currentStore) {
+    currentStore = createStoreInstance();
+  }
+
+  return currentStore;
+};
+
 // this needs to be run last
 export default createRuntimePluginAfter(
   {
+    appContext: ctx => {
+      let initialState = {};
+      if (!isServer) {
+        initialState = getPageData('shuviInitialState', {});
+      }
+      ctx.store = initStore({
+        ctx,
+        initialState
+      });
+    },
     appComponent: async (App, appContext) => {
       return DouraWrapper(App, {
         store: appContext.store
